Checkpointing for the evening: needs rework around unit tests for xmlutils.
